+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added support for petabytes and exabytes
 - Fixed error with `xmin` and `xmax` and empty data
+- Fixed error with `GeoChart` with latest Google Charts release
 
 ## 3.2.0 (2019-11-09)